Update 12-22-2024
Fixed the Explosive Barrel script. Applied changes in the Targeter Sound Binding script.

Tried removing the +ISMONSTER flag but, when doing so, the Targeter will not acknowledge the barrel. Not sure as to why since it's able to register the Pod and Poison Mushroom, however, I'll see if I can approach this issue from a different angle.